// OpenWMS-new/src/app/common/model/user.model.ts
export class UserModel {
    user_code: string;
    user_id: string;
    user_name: string;
    user_password: string;
    name: any;
    user_image: string;
    user_email: string;
    phone: string;
    department_id: string;
    user_sex: string;
    user_age: string;
    user_card: string;
    user_birthday: string;
    unit_id: string;
    is_enable: string;
    open_id: string;
    creator_id: string;
    create_date: Date;
    modifier_id: string;
    role_name: string;
    login_time: Date;
    site: string;
    desc: string;
    modify_date: Date;
}