// OpenWMS-new/src/app/common/model/match.mode.ts
export class MatchModel {
  match_id: string;
  match_name: string;
  match_start_date: Date;
  match_end_date: Date;
  match_type: string;
  match_writing: string;
  match_num: string;
  match_state: string;
  enroll_num: string;
  enlist_start_date: Date;
  enlist_end_date: Date;
  enlist_mode: string;
  enlis_phone: string;
  enlis_person: string;
  enlis_picture: string;
  enlis_big_picture: string;
  enlis_price: string;
  Is_enable: string;
  creator_id: string;
  create_date: Date;
  modifier_id: string;
  modify_date: Date;
}